requirements.txt
django-pain>=0.2.0
django-app-settings
fred-pyfco
